Continue post-merge port: reth-mptdb compiles end-to-end (dev mode runs)
Final round of conflict-resolution after the dev-v2.2.0 merge so the
reth-mptdb dev node boots and produces canonical blocks.

Key changes:
- node/builder/rpc.rs: extend Debug impl and NodeAddOns impl on
  RpcAddOns with the AuthHttpMiddleware 7th generic and
  Send + Sync + 'static bounds.
- ethereum/node: tighten EthereumPayloadBuilder / components_builder
  bounds with PayloadBuilderAttributes = EthPayloadBuilderAttributes;
  add Send + Sync + 'static bound for AuthHttpMiddleware on
  NodeAddOns / RethRpcAddOns / EngineValidatorAddOn impls of
  EthereumAddOns.
- bin/reth-mptdb: switch BundleState init to AddressMap / B256Map to
  match revm-database 13's hashed maps; add the new
  write_storage_changesets field to StateWriteConfig.
- engine/service: restore HEAD's EngineService crate around the new
  reth_tasks::Runtime spawner.
- engine/tree, engine/primitives, engine/util, engine/local,
  payload/basic, payload-builder, ethereum/payload, ethereum/engine-
  primitives, rpc, rpc-engine-api, rpc-convert, debug-client, evm,
  stages, storage/provider, tasks, trie/db, chain-state,
  primitives-traits, node/{api,builder,core}: misc post-merge fixups
  (TaskSpawner alias, ExecutionPayloadEnvelopeV6 alias, FastInstant
  signatures, version-arg on fork_choice_updated, restored HEAD
  versions of the items reth-mptdb depends on, etc.).

Smoke test: dev node opens DB, pre-populates genesis into mptdb,
starts RPC + auth servers, and produces blocks 1-N at 1s cadence
with no errors.

+++ crates/node/core/src/args/engine.rs
@@ -645,7 +645,7 @@ impl EngineArgs {
             .without_prewarming(self.prewarming_disabled)
             .with_state_provider_metrics(self.state_provider_metrics)
             .with_always_compare_trie_updates(self.state_root_task_compare_updates)
-            .with_cross_block_cache_size(self.cross_block_cache_size * 1024 * 1024)
+            .with_cross_block_cache_size(self.cross_block_cache_size as u64 * 1024 * 1024)
             .with_multiproof_chunk_size(self.multiproof_chunk_size)
             .with_reserved_cpu_cores(self.reserved_cpu_cores)
             .without_precompile_cache(self.precompile_cache_disabled)
